Loading files now updates panel text
